Apply the patch for the async signals race condition in Sendmail.

+++ lib/libssl/src/crypto/dh/dh.h
@@ -101,7 +101,7 @@ struct dh_st
 	int version;
 	BIGNUM *p;
 	BIGNUM *g;
-	long length; /* optional */
+	int length; /* optional */
 	BIGNUM *pub_key;	/* g^x */
 	BIGNUM *priv_key;	/* x */
